requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
autopep8
pylint
pytest
iso-language-codes